refactor: inputView, outputView용 별도 View 패키지 생성
